# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateStoreRequest:
    properties:
      name:
        type: string
      owner_id:
        type: string
    required:
    - name
    type: object
  dto.CreateTenantRequest:
    properties:
      name:
        type: string
      store_id:
        type: string
    required:
    - name
    - store_id
    type: object
...
  dto.StoreResponse:
    properties:
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
  dto.TenantResponse:
    properties:
      id:
        type: string
      name:
        type: string
      store_id:
        type: string
    type: object
...
  dto.UpdateStoreRequest:
    properties:
      name:
        type: string
      owner_id:
        type: string
    required:
    - name
    type: object
  dto.UpdateTenantRequest:
    properties:
      name:
        type: string
      store_id:
        type: string
    required:
    - name
    - store_id
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Point of Sale (POS) application with multi-tenant support.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: POS Backend API
  version: "1.0"
paths:
...
  /stores:
    get:
      description: Retrieve list of stores (optionally filtered by owner_id)
      parameters:
      - description: Owner ID
        in: query
        name: owner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StoreResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all stores
      tags:
      - Store
  /stores/:
    post:
      consumes:
      - application/json
      description: Create a new store. If owner_id is omitted, it's taken from the
        authenticated user.
      parameters:
      - description: Create Store Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create store
      tags:
      - Store
  /stores/{id}:
    delete:
      description: Delete a store by ID
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete store
      tags:
      - Store
    get:
      description: Retrieve a single store by its ID
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get store by ID
      tags:
      - Store
    put:
      consumes:
      - application/json
      description: Update an existing store by ID
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Store Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update store
      tags:
      - Store
  /stores/filter:
    get:
      description: Retrieve stores using name filter, pagination and owner filter
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Owner ID
        in: query
        name: owner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StoreResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get stores with filter and pagination
      tags:
      - Store
  /tenants:
    get:
      description: Retrieve list of tenants (optionally filtered by store_id)
      parameters:
      - description: Store ID
        in: query
        name: store_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TenantResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all tenants
      tags:
      - Tenant
  /tenants/:
    post:
      consumes:
      - application/json
      description: Create a new tenant. `store_id` must reference an existing store.
      parameters:
      - description: Create Tenant Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TenantResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create tenant
      tags:
      - Tenant
  /tenants/{id}:
    delete:
      description: Delete a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete tenant
      tags:
      - Tenant
    get:
      description: Retrieve a single tenant by its ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TenantResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get tenant by ID
      tags:
      - Tenant
    put:
      consumes:
      - application/json
      description: Update an existing tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Tenant Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TenantResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update tenant
      tags:
      - Tenant
  /tenants/filter:
    get:
      description: Retrieve tenants using name filter, pagination and store filter
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Store ID
        in: query
        name: store_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TenantResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get tenants with filter and pagination
      tags:
      - Tenant
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"